+++ magicasper/public/index.html
@@ -7,6 +7,7 @@
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
+    <div id="root">
+    </div>
   </body>
 </html>
