--- magicasper/src/Componentes/Disciplinas/index.js
@@ -1,9 +1,0 @@
-import './index.css';
-
-function Disciplinas (){
-return <h1>Disciplinas</h1>
-
-};
-
-
-export default Disciplinas
