+++ magicasper/src/Componentes/Navbar/index.js
@@ -1,5 +1,6 @@
 import './index.css';
 import { Link } from "react-router-dom"
+import {MdOutlineSettings} from 'react-icons/md'
 
 function Navbar  () {
 return <header>
@@ -12,7 +13,7 @@ return <header>
               <li><Link to="/Home">Home</Link></li>
               <li><Link to="/Perfil">Perfil</Link></li>
               <li><Link to="/Disciplinas">Materias</Link></li>
-              <li><img src="./img/config-icon.png" alt="Config Icon" /></li>
+              <li className='setting-icon'><MdOutlineSettings/></li>
             </ul>
           </div>
         </nav>
