+++ magicasper/src/Componentes/mainLabel/index.css
@@ -1,0 +1,27 @@
+#Conteudo-Disciplinas li{
+  list-style: none;
+  float: left;
+  padding-top: 10px;
+}
+
+#Conteudo-Disciplinas .icone-selecao{
+  width: 20px;
+  height: 31.03px;
+  flex-shrink: 0;
+  margin-left: 36px;
+  padding-top: 12px;
+}
+
+#Conteudo-Disciplinas label{
+  color: #FFF;
+  text-align: center;
+  font-family: Reem Kufi;
+  font-size: 20px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: normal;
+  cursor: pointer;
+}
+.icone-selecao svg{
+  cursor: pointer;
+}
